+++ src/main/java/com/matdang/seatdang/waiting/controller/dto/WaitingRequest.java
@@ -8,5 +8,5 @@ import lombok.Data;
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 public class WaitingRequest {
     private Long storeId;
-    private Long peopleCount;
+    private Integer peopleCount;
 }
